requirements.txt
Django==4.2.7
Pillow==10.1.0
django-jazzmin==2.6.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-environ==0.11.2
whitenoise==6.6.0
gunicorn==21.2.0
Python==3.13.7